requirements.txt
# dlib
numpy
opencv-python-headless
Flask
Pillow
gunicorn
retinaface
tf-keras 